backend/requirements.txt
# Core FastAPI
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
supabase>=2.3.5

# AI/LLM
openai>=1.12.0
anthropic>=0.28.0

# Google APIs
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.107.0
google-auth>=2.25.2

# HTTP clients
httpx>=0.25.2
aiohttp>=3.9.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
